pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "databricks-mlops"
version = "0.1.0"
description = "A comprehensive framework for MLOps on Databricks"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.5.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "databricks-sdk>=0.20.0",
    "mlflow>=2.9.0",
    "great-expectations>=0.18.0",
    "delta-spark>=2.4.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "logging-config>=1.1.0",
    "pyyaml>=6.0.1",
    "pytest>=7.4.3",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.10.0",
    "isort>=5.12.0",
    "mypy>=1.6.1",
    "ruff>=0.1.4",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.5.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/databricks-mlops"
"Bug Tracker" = "https://github.com/yourusername/databricks-mlops/issues"

[tool.hatch.build.targets.wheel]
packages = ["databricks_mlops"]

[tool.uv]
repository = "https://pypi.org/simple"
export = ["requirements.txt", "requirements-dev.txt"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
